TS: split RelatedSimple from components/Related
